requirements.txt
pytest==8.4.1
requests==2.31.0
allure-pytest==2.14.2
python-dotenv==1.0.0 